refactor: Merge routine, exercise-to-routine dto files
