--- packages/backend/src/routines/dto/create-routine.dto.ts
@@ -1,3 +1,0 @@
-export class CreateRoutineDTO {
-  name: string;
-}
