--- packages/backend/src/routines/dto/remove-exercise-from-routine.dto.ts
@@ -1,3 +1,0 @@
-export class RemoveExerciseFromRoutineDTO {
-  exerciseId: string;
-}
