--- packages/backend/src/routines/dto/update-routine.dto.ts
@@ -1,3 +1,0 @@
-export class UpdateRoutineDTO {
-  name: string;
-}
